+++ NEWS.md
@@ -9,7 +9,7 @@
 * Provides `item-storage-batch-sync-unsafe 2.0`
 * Provides `item-storage-dereferenced 1.2`
 * Provides `holdings-storage 8.1`
-* Provides `instance-storage 11.2`
+* Provides `instance-storage 11.3`
 * Provides `service-points 3.5`
 * Provides `inventory-hierarchy 0.8`
 * Provides `inventory-view 3.2`
@@ -44,6 +44,7 @@
 * Add default empty values for S3_SECRET_ACCESS_KEY and S3_ACCESS_KEY_ID ([MODINVSTOR-1489](https://folio-org.atlassian.net/browse/MODINVSTOR-1489))
 * Implement sanitizers for HoldingsRecord, Instance, and Item ([MODINVSTOR-1497](https://folio-org.atlassian.net/browse/MODINVSTOR-1497))
 * Create a submodule with generated DTOs ([MODINVSTOR-1499](https://folio-org.atlassian.net/browse/MODINVSTOR-1499))
+* Implement of PATCH requests for instances ([MODINVSTOR-1456](https://folio-org.atlassian.net/browse/MODINVSTOR-1456))
 
 ### Bug fixes
 * Fix ordering of electronic access items for inventory-hierarchy, oai-pmh-view ([MODINVSTOR-1224](https://folio-org.atlassian.net/browse/MODINVSTOR-1224))